--- a/packages/client/src/index.ts
+++ b/packages/client/src/index.ts
@@ -16,4 +16,4 @@ export * from './errors';
 export * from './workflow-options';
 export * from './interceptors';
 
-export { DataConverter, defaultDataConverter } from '@temporalio/workflow';
+export { DataConverter, defaultDataConverter } from '@temporalio/common';