--- a/packages/common/src/errors.ts
+++ b/packages/common/src/errors.ts
@@ -1,5 +1,20 @@
 import { coresdk } from '@temporalio/proto';
 
+export class ValueError extends Error {
+  public readonly name: string = 'ValueError';
+}
+
+export class DataConverterError extends Error {
+  public readonly name: string = 'DataConverterError';
+}
+
+/**
+ * Used in different parts of the project to signal that something unexpected has happened
+ */
+export class IllegalStateError extends Error {
+  public readonly name: string = 'IllegalStateError';
+}
+
 export function errorToUserCodeFailure(err: unknown, nonRetryable?: boolean): coresdk.common.IUserCodeFailure {
   if (typeof err === 'string') {
     return { message: err };