--- a/packages/common/src/index.ts
+++ b/packages/common/src/index.ts
@@ -1,0 +1,12 @@
+/**
+ * Common library for both isolated Workflow and normal non-Workflow code
+ *
+ * @module
+ */
+export * from './errors';
+export * from './interfaces';
+export * from './activity-options';
+export * from './converter/data-converter';
+export * from './interceptors';
+export * from './time';
+export * from './tls-config';