--- a/packages/common/src/index.ts
+++ b/packages/common/src/index.ts
@@ -10,3 +10,5 @@ export * from './converter/data-converter';
 export * from './interceptors';
 export * from './time';
 export * from './tls-config';
+// NOTE: workflow-options not exported because they rely on upstream proto
+// This dependency will be removed soon and the export will be added.