--- a/packages/common/src/interceptors.ts
+++ b/packages/common/src/interceptors.ts
@@ -1,4 +1,4 @@
-import { coresdk } from '@temporalio/proto';
+import { coresdk } from '@temporalio/proto/lib/coresdk';
 import { AnyFunc, OmitLastParam } from './type-helpers';
 
 /**