--- a/packages/common/src/time.ts
+++ b/packages/common/src/time.ts
@@ -1,6 +1,6 @@
 import Long from 'long';
 import ms from 'ms';
-import * as iface from '@temporalio/proto';
+import * as iface from '@temporalio/proto/lib/coresdk';
 import { ValueError } from './errors';
 
 // NOTE: these are the same interface in JS