--- a/packages/proto/lib/index.d.ts
+++ b/packages/proto/lib/index.d.ts
@@ -22,3 +22,6 @@
  * ```
  * @module
  */
+
+export * from './temporal';
+export { coresdk } from './coresdk';