--- a/packages/proto/lib/index.js
+++ b/packages/proto/lib/index.js
@@ -1,0 +1,2 @@
+module.exports = require('./temporal');
+module.exports.coresdk = require('./coresdk').coresdk;