Closes #100

I did this now because I need to share workflow-client between workflow and client packages for ChildWorkflow.