@bergundy 2021-08-02
Looks like performance improved.
Results from single round of perf tests that I ran on my M1 MBP.

```shellsession
$ time node packages/test/lib/perf/all-in-one.js --iterations 30000 --workflow sync --concurrent-wf-clients 500 --max-cached-wfs 2500
```

Before change:
```
WFs complete (30000/30000), WFs/s curr 9.6 (acc 48.1)
642.31s user 103.73s system 119% cpu 10:25.78 total
```

After change:
```
WFs complete (30000/30000), WFs/s curr 9.1 (acc 66.2)
481.74s user 30.10s system 111% cpu 7:40.02 total
```