@bergundy 2021-08-03
I haven't evaluated separate node processes but I have a feeling it'll be much more expensive.
The only "valid" options I see right now are:
1. Staying with isolates for the WF runtime.
2. Running WFs in the default node isolate (in the same process as the Worker)

Going with (2) means that we lose all of the work we've put in to create a deterministic runtime for WFs and going with (1)  adds a lot of work (mostly done already) and a performance hit to the SDK which hopefully is either acceptable or can be improved to be acceptable - ~70 WF/s on my computer using up just over 100% CPU.

The first thing we did when working on this SDK is build the deterministic WF isolate environment, we've already invested a lot of energy in this path and I hope it'll pay off.

A subprocess wouldn't help either of the approaches, it'll just add overhead.
FTR Each node process takes about 30 MB of RAM just to start up and about 50ms to start on my laptop.